Listo para pruebas de campo

+++ Botella.h
@@ -5,13 +5,13 @@
 
 typedef struct {
   uint16_t capacidad = 330;          // (cc) Capacidad de la botella.
-  uint16_t tMovLlenador = 3;         // (Seg.) Espera movimiento llenador.
-  uint16_t tMovTapador = 3;          // (Seg.) Espera movimiento Tapador.
-  uint16_t tCargaCO2_Barrido = 2;    // (Seg.) carga CO2 Barrido inicial.
-  uint16_t tDescargaCO2_Barrido = 1; // (Seg.) descarga CO2 Barrido incial.
-  uint16_t tCargaBeer = 10;          // (Seg.) carga de cerveza.
-  uint16_t tEstBeer = 2;             // (Seg.) Estabilizacion presion beer.
-  uint16_t tEstCO2 = 2;              // (Seg.) Estabilizacion presion CO2.
+  byte tMovLlenador = 3;         // (Seg.) Espera movimiento llenador.
+  byte tMovTapador = 3;          // (Seg.) Espera movimiento Tapador.
+  byte tCargaCO2_Barrido = 2;    // (Seg.) carga CO2 Barrido inicial.
+  byte tDescargaCO2_Barrido = 1; // (Seg.) descarga CO2 Barrido incial.
+  byte tCargaBeer = 10;          // (Seg.) carga de cerveza.
+  byte tEstBeer = 2;             // (Seg.) Estabilizacion presion beer.
+  byte tEstCO2 = 2;              // (Seg.) Estabilizacion presion CO2.
 } TBotella;
 #define botellasTam 6
 TBotella botellas[botellasTam];
